Essential process fixes
